Add cluster imputation in assignClusters()
Add cluster imputation in assignClusters() with corresponding
arguments `redo_imputation` and `knn_imp`.
The imputation should be rerun at the assignClusters() if the clusters
are updated (custom clusters or remapping) to ensures the imputed
clusters are updated.
Imputation can be avoided by setting `redo_imputation` to FALSE
(TRUE by default).
The `clusters` argument can contain only common cells if
`redo_imputation` is TRUE, as the imputation will be rerun.