Update and refactor cnaCalling()
- Update to new format : now the allelic counts of normal cells
are not aggregated on all omics beforehand (in aggregateCounts()
the counts are seperated by omics) so they are aggregated in the
same way as tumor counts for allelic.

- Refactor processing of counts and filtering : separation of
allelic and coverage counts for (1) aggregating allelic counts
by position for all omics, (2) optionally selecting omics for
coverage, and (3) applying filters on both separately. Both for
"per cluster" and "all cells" processing.

- Fix for clusters in character format
(expected output of aggregateCounts() in combined.counts table)